Analytics Controller & Services, DTOs and updated postman .json and added more mock data generated by ClaudeAI

+++ src/main/java/at/qe/skeleton/dtos/DepartmentAnalyticsDTO.java
@@ -9,5 +9,5 @@ public record DepartmentAnalyticsDTO(
         String departmentName,
         int roomCount,
         int activeViolations,
-        Map<Metric, Long> avgMetrics
+        Map<Metric, Double> avgMetrics
 ) {}
