+++ src/main/java/at/qe/skeleton/dtos/TrendPointDTO.java
@@ -4,5 +4,5 @@ import java.time.LocalDateTime;
 
 public record TrendPointDTO(
         LocalDateTime timestamp,
-        Long value
+        Double value
 ) {}
